app/requirements.txt
# D:\Docker\BrokerSummary\app\requirements.txt
streamlit
pandas
mysql-connector-python
matplotlib
seaborn
yfinance
plotly
ta
psycopg2-binary